requirements.txt
altair==5.0.1
anyio==3.7.1
arrow==1.2.3
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
boto3==1.27.0
botocore==1.30.0
bravado==11.0.3
bravado-core==5.17.1
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.99.1
fqdn==1.5.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.5.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
isoduration==20.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.17.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
msgpack==1.0.5
neptune==1.3.2
numpy==1.24.4
oauthlib==3.2.2
orjson==3.9.1
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
psycopg2-binary==2.9.6
pyarrow==12.0.1
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.4.2
s3transfer==0.6.1
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
streamlit==1.24.0
streamlit-aggrid==0.3.4.post3
swagger-spec-validator==3.0.3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
webcolors==1.13
websocket-client==1.6.1
websockets==11.0.3
zipp==3.15.0
